move logs + logger.js
